kontron_mx6ul: Use the SPI NOR as a fallback boot device to load U-Boot proper
The default boot fuse settings use the SD card (MMC1) as primary boot
device, but allow SPI NOR as a fallback boot device. We cannot use the
SPI NOR as primary boot otherwise we may mix the SPL (from the MMC1) and
U-Boot proper from the SPI NOR. Also if no U-Boot image is flashed to
SPI NOR, the boot will be interrupted as the SPI NOR access is
blocking the second boot device (MMC).

If booting for MMC we should use the U-Boot proper from MMC as first
try and SPI NOR only second.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>
[Adjust comment/commit message and refactor code]
Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>