imx8m: Fix DDR runtime size detection with two banks
If we want to detect DDR sizes up to 8G at runtime, we need to set
PHYS_SDRAM_SIZE and PHYS_SDRAM_2_SIZE to their maximums of 3G and
5G respectively. In case there is actually only 4G of RAM available,
we need to calculate the correct sizes for bank 1 and 2.

This method works if board_phys_sdram_size() is defined to return
the actual RAM size accross both banks by calling get_ram_size().

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>